CurrencyPriceAndCompareTexts done
